project structure refactor
